cross product and handness clarification, should polish

+++ src/os_win32.h
@@ -14,7 +14,7 @@ struct Software_Depth_Buffer
     i32 width;
     i32 height;
     BITMAPINFO info;
-    u32 *data;
+    f32 *data;
 };
 
 typedef struct OS_W32_Window OS_W32_Window;
